Added new SqlStorage implementation as an abstract base class for all RDBMS storage implementations
The base class utilizes SqlAlchemy for connection management and SQL statement execution because it is the only connection pooling library that works with the Snowflake connector.

The base class includes the background db write capabilities from the Snowflake implementation and actual SQL statement execution where standard SQL is used.  Abstract methods are defined for queries where the SQL or database APIs are not standard.