fix: remove `search_variations` and `get_variation_count` declarations from the `_Storage` class (#237)
In Issue #228, I added abstract declarations for `search_variations` and
`get_variation_count` declarations in the `_Storage` class, which is the
abstract base parent class of all other storage classes in AnyVar. I
forgot that the Annotation Storage classes _also_ inherit from this, but
they of course do not implement these methods. That _\*ahem\*_ may have
broken things. This PR reverts that change so that AnyVar works again.

closes ##238