feat: allow table name overrrides (issue #252)
closes #252 

## Pluralization 

Note that I maintained our convention of utilizing pluralized table
names by simply appending an "s" to the end of the ORM class names when
creating the tables. This will obviously not work in the future if we
decide to add classes that are pluralized irregularly (e.g., a class
named `Mouse` would produce a table named `mouses`, which is of course
incorrect).

There are libraries that can handle this problem for us, but since our
current tables don't require this solve, I opted for this simple
solution instead. However, if we think it wise to make our pluralization
more robust now so that we're protected in the future, I can change
this.

## About `@declared_attr.directive`

See the SqlAlchemy documentation on this
[here](https://docs.sqlalchemy.org/en/20/orm/mapping_api.html#sqlalchemy.orm.declared_attr)
if you're curious