chore: add VCF benchmarking script (#279)
close #275 

Well, I'm not sure if this closes 275, it depends on what we were
thinking for that issue. But this is a thin rewrite of a script I wrote
up for GREGoR that can be used to benchmark VCF submissions. It can take
multiple input files as arguments and progressively submit them/monitor
their progress against a running AnyVar API. There's a multiprocessing
part that's there not for benchmarking per se but as a convenience for
adding a bunch of big VCFs (eg GREGoR).

Initial unscientific testing on my laptop indicated that the current
`storage-refactor-epic` commit performed about as well as `main` on a
small (10k records) VCF

## Questions
* Currently polls the status endpoint every 5 seconds. That might be too
long