feat!: use updated `agct` release (#285)
The new agct release includes several changes that make our lives easier
over here:
* A cached `get_converter()` function, so that we don't need to keep
track of existing liftover machine instances ourselves
* Functions to go back and forth between a refget accession ID and the
corresponding reference assembly and chromosome values, meaning we don't
need to use SeqRepo in the liftover part (it is still needed to
normalize the output allele)
* This simplifies dependencies for test cases, enabling some of the
cases to very easily run in github actions. It does technically change a
public API though, which is why this is a breaking change
* Liftover can be performed on an interval, not just a position. I think
this is ideal, but thinking about how to implement this for the typing
insanity of the SequenceLocation start and end parameters was making my
head spin, so I left it as a todo.

This PR bumps the dependency up to `>= 0.2.0rc1`. Eventually that could
be bumped to just `0.2.0` if needed, once all `agct` downstream users
have implemented the required changes.