chore: add more granular logging to queueing dependency guards (#299)
I was running some celery-based vcf annotation tests yesterday and
having an issue where one of my queuing dependencies wasn't working
(separate dumb mistake). Unfortunately, it was kind of tricky to
diagnose where and why it was failing. I added these possibly
overly-detailed logging statements to show a) the values of each
parameter so that you can tell why it's failing and b) the stack trace
so you can see where it's failing.