feat: use batch db writes for vcf variant registration (issue #253)
This PR does the following:

1. Renames the `anyvar.put_object` method to `anyvar.put_objects`
(plural) and updates it to accept an array of objects instead of just a
single object (in line with our implementation of our underlying storage
method, `add_objects`).
2. Makes the VCF registration endpoints (`PUT /vcf` and `PUT
/annotated_vcf`) use batch database writes. I used a batch size of
`10,000`, but we can change this if we want.

I tested this using a ~800MB VCF file with 5.7 million variants. I ran
it through the `PUT /vcf` endpoint on both `main` and this feature
branch, and the runtimes were _slightly_ slower on this branch but
overall seem fairly comparable:

| | `main` | This branch | Difference vs `main` |
|-|--------|-------------|----------------------------------|
| Runtime (minutes) | 317.12 minutes| 329.95 minutes | +12.83 minutes |
| Runtime (hours) | 5.29 hours | 5.50 hours | +0.21 hours |

I can experiment more with the batch size if we think it could have a
meaningful impact on runtimes! Reviewers, I'd love your thoughts on
this.

closes #253