feat: solidify storage interface edge cases/error handling (#271)
* Fill out docstrings for storage class methods. Leave a note in
`NoObjectStore` methods that they don't do anything
* Raise a specific error in `add_objects` if VRS object properties are
missing or if objects aren't fully dereffed
* No `ONDELETE CASCADE` anywhere
* Raise an exception when trying to delete a thing that is depended upon
by another thing (e.g. delete a `SequenceLocation` that is referenced by
an `Allele`)
* Define an exception when trying to add an annotation or mapping if the
referenced variation isn't in the DB (left as a TODO for #286 )
* Make type filter optional in `get_mappings`, in order to match
behavior in `get_annotations`
* `add_annotation` returns `None` to match `add_mapping`
* `add_annotation` explicitly creates redundant rows (can't remember if
we decided this was the intended behavior or not)
* `delete_annotation` takes an `Annotation` object as an arg rather than
a DB ID (which is more ephemeral)
* `Annotation` and `VariationMapping` domain classes don't include the
database ID
* `Annotation.annotation_value` is a `JsonValue`, not an `Any`.
Otherwise, we may be trying to shove non-serializable things into
Postgres (sometimes this is fine, sometimes this works but produces a
lossy change, othertimes it'll raise an error from postgres --
`JsonValue` just checks to ensure a value is "json-able"). This could
also be changed in `anyvar.restapi.schema.AddAnnotationRequest` but I'm
trying to remain scoped to the storage class
* Raise an error in `search_alleles` for invalid start/end positions
* Leave some TODO's in `search_alleles` for behavior that we've
discussed but haven't come to a conclusion about
* Add tests for all of the above as needed