fix: return 404 when resource not found (#322)
close #304 

Return a 404 if the VRS ID in `/variation/<vrs_id>/annotations` or
`/variations/<vrs_id>/mappings` doesn't exist.

If the `ObjectStorage.get_annotations()` lookup returns an empty list,
*then* do a check to see if the VRS ID exists at all. If it doesn't,
then return the error/a 404.

### Questions/thoughts

* It feels weird to do this for mappings ONLY if the source ID doesn't
exist (not the destination ID), but IMO that seems like the most RESTy
way. If instead the path was something like `GET
/variation/<source_id>/mappings/<dest_id>/` then I think I'd be more
comfortable trying to do this for both.