refactor: streamline supported vrs object types (issue #265)
closes #265 

### Overview

I streamlined our `VrsObject` and `VrsVariation` types and related enums
so that we have one source of truth for each.

The biggest updates in this PR are in `src/anyvar/utils/types.py`: 
* I got rid of the `StoredObjectType` enum so that we're just using the
`VrsObject` type union instead.
* I wanted to get rid of our `SupportedVariationType` enum, but we
really do need this for our `RegisterVariationRequest` model; so instead
I am creating the enum dynamically based on the models listed in our
`VrsVariation` type union
* I also updated our `vrs_object_class_map` to be generated dynamically
so we don't need to manually add things to it each time we want to
support a new type of Vrs Object.

### TLDR:
Previously, if we were to add support for a new type of VRS Variation,
we'd have to add it in 5 places:
* The `StoredObjectType` enum
* The `VrsObject` type union
* The `VrsVariation` type union
* The `SupportedVariationType` enum
* The `variation_class_map` dict

Now, we'll only have to update 2 places:
* The `VrsObject` type union
* The `VrsVariation` type union