docs: rewrite everything (#317)
close #208 
close #297 

* Sprawling docs rewrite. Remove a lot of old or less-necessary stuff,
reorganize globally. Previously we just had a set of separately-written
single-use pages; this PR tries to synthesize something more
comprehensive. Somewhat vaguely trying to follow the tenets of
https://diataxis.fr/ on a page-by-page basis
* Add features section. This is pretty sparse now but should be filled
in with high-level "what is AnyVar for, what can it do"-type
descriptions that subsequent sections in the docs refer back to.
* Add "usage" sections
    * Using it in Python apps/consoles
    * Using the REST API
    * Running the server
* Clean up README -- in particular, remove redundant or repetitive stuff
and direct users to docs instead

I think there are still a lot of little notes that could be added which
I may be forgetting.

Questions

* I don't ever use the Docker Compose file, but I see that it seems to
contain a lot of optional-ish things depending on whether you want to
reuse resources. Is the provided file good as-is for `docker-compose up`
or is there additional configuration needed? I thought about trying to
fill out a configuration page just for this but didn't really know what
I was looking at. (wound up filling in some stuff)
* Mention `SEQREPO_ROOT_DIR` somewhere?
* mention dev team/"who we are"/organizational backing somewhere?

Possible future directions

* Fill in stuff for VLM as needed
* Flesh out "features" more
* Add something more in the vein of "recipes" for demonstrating
complicated tasks
* add sphinx doctest